+++ plugin/dubbo/route.go
@@ -1,4 +1,4 @@
-package generator
+package dubbo
 
 import (
 	"fmt"
